fastAPI-for-data-science/04_observability/pyproject.toml
[project]
name = "04-observability"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.2",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pandas>=3.0.1",
    "ruff>=0.15.4",
    "scikit-learn>=1.8.0",
    "skl2onnx>=1.20.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
